[BACK-END REFACTOR] made different files for each model of the schema
