+++ Back-end/schema/models/Car.js
@@ -1,0 +1,9 @@
+const mongoose = require('mongoose');
+const Dimensions = require('./Dimensions');
+
+const Car = mongoose.Schema({
+    dimensions: Dimensions,
+    plates: String
+}, { _id: false });
+
+module.exports = Car;
