+++ Back-end/schema/models/Dimensions.js
@@ -1,0 +1,9 @@
+const mongoose = require('mongoose');
+
+const Dimensions = mongoose.Schema({
+    heightFts: Number,
+    lengthFts: Number,
+    widthFts: Number,
+}, { _id: false });
+
+module.exports = Dimensions;
